+++ tpl/js/main.js
@@ -8,12 +8,21 @@
 
 	/////////////////setup/////////////////
 	app.setup({
-		theme: 'project',
 		template: '@{{{template}}}',
-		navRegion: 'content' 
+		navRegion: 'content',
 
 		//Note: Always set navRegion if using app template here, unless you've merged it(the tpl) with index.html;
 		
+		defaultContext: 'Default',
+		viewSrcs: 'js', //set this to a folder path to enable view dynamic loading. 
+		//---------------------------------------------------------------------------------------------
+		fullScreen: false, //this will put <body> to be full screen sized (window.innerHeight).
+		//---------------------------------------------------------------------------------------------
+		i18nTransFile: 'i18n.json', //can be {locale}.json
+		i18nLocale: '', //if you really want to force the app to certain locale other than browser preference. (Still override-able by ?locale=.. in url)
+		//---------------------------------------------------------------------------------------------
+		baseAjaxURI: '', //modify this to fit your own backend apis. e.g index.php?q= or '/api'
+		timeout: 5 * 60 * 1000 //general communication timeout (ms), e.g when using app.remote()
 	});
 
 	///////bootstrapping events//////////// - [optional]
@@ -44,8 +53,5 @@
 		//your engine apis for managing this app's state. (e.g invoked in Contexts and Views)
 
 	});
-
-	///////////////kick start//////////////
-	app.run(/*'deviceready' if you are developing with Cordova*/);
 
 })(Application);