+++ tpl/js/router.js
@@ -5,12 +5,14 @@
  * ----
  * Routers are automatically loaded by the devserver after creation.
  *
- * Example
- * -------
+ * Default mount path (by filename)
+ * ------------------
  * /abc => routers/abc.js
  * /abc/efg => routers/abc/efg.js
  * /sample/sample2-abc => routers/sample/sample2Abc.js 
  * /sample/smaple2_abc => routers/sample/sample2Abc.js
+ *
+ * Use server.mount(this, custom-url) to use your own url instead of the default mount paths.
  *
  * Warning
  * -------
@@ -30,14 +32,14 @@
 module.exports = function(server){
 
 	var profile = server.get('profile');
-	var router = server.mount(this);
-	//server.secure(router, /*token1, token2, ...*/); -- see util/secure.js
+	var router = server.mount(this /*, custom url*/);
+	//server.secure(router, interpretations); -- see util/secure.js
+	//						interpretations defines { 'token': fn(req){ return true/false; }, ...} customized checking.
+
+	//////////routes, +optional permission token checking in req.session//////////
 	
-	////////////////////////////////////////////
-
-
 	//get '/'
-	router.get('/', /*router.token(t1, t2, ...)*/ function(req, res, next){
+	router.get('/', /*router.token(t1, t2, ...),*/ function(req, res, next){
 		res.json({stagejs: '!', generated: '{{date}}'});
 	});